backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
psutil>=5.9.0
python-multipart>=0.0.9